backend/requirements.txt
# AI 케어브릿지 Backend Dependencies
# Python 3.11+

# === Core ===
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# === LangChain & LangGraph ===
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-upstage>=0.1.0
langgraph>=0.0.26

# === Vector DB & Embeddings ===
chromadb==0.4.22
sentence-transformers==2.2.2

# === Memory & Cache ===
redis==5.0.1
aioredis==2.0.1

# === Speech ===
openai-whisper==20231117
openai==1.10.0
edge-tts==6.1.12

# === Utils ===
httpx==0.26.0
python-multipart==0.0.21
aiofiles==23.2.1

# === Testing ===
pytest==8.0.0
pytest-asyncio==0.23.0
